Random forest code; change filenames
"train_rf" added to /lib/. This includes code that performs cross validation to select the best parameters for the random forest model. It also builds the best such model on our basic train/test split, and records the amount of time necessary and accuracy that results. Also, I changed some filenames in our overall directory structure so we know what we have to update or remove to make a clean final deliverable.